moved annotation to java package
