+++ actor-testkit-typed/src/test/java/jdocs/org/apache/pekko/actor/testkit/typed/javadsl/Junit5IntegrationExampleTest.java
@@ -11,6 +11,7 @@
 
 package jdocs.org.apache.pekko.actor.testkit.typed.javadsl;
 
+import org.apache.pekko.actor.testkit.typed.annotations.Junit5TestKit;
 import org.apache.pekko.actor.Address;
 import org.apache.pekko.actor.testkit.typed.javadsl.*;
 import org.apache.pekko.actor.testkit.typed.javadsl.Junit5TestKitBuilder;
@@ -27,7 +28,8 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 @ExtendWith(TestKitJunit5Extension.class)
 class Junit5IntegrationExampleTest {
 
-  @Junit5TestKit public ActorTestKit testKit = new Junit5TestKitBuilder().build();
+  @Junit5TestKit
+  public ActorTestKit testKit = new Junit5TestKitBuilder().build();
 
   @Test
   void junit5Test() {
