+++ actor-testkit-typed/src/test/java/org/apache/pekko/actor/testkit/typed/javadsl/ActorTestKitJunit5Test.java
@@ -9,6 +9,7 @@
 
 package org.apache.pekko.actor.testkit.typed.javadsl;
 
+import org.apache.pekko.actor.testkit.typed.annotations.Junit5TestKit;
 import org.apache.pekko.Done;
 import org.apache.pekko.actor.typed.javadsl.Behaviors;
 
@@ -30,7 +31,8 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 @ExtendWith(LogCapturingExtension.class)
 class ActorTestKitJunit5Test extends JUnitSuite {
 
-  @Junit5TestKit public ActorTestKit testKit = new Junit5TestKitBuilder().build();
+  @Junit5TestKit
+  public ActorTestKit testKit = new Junit5TestKitBuilder().build();
 
   @Test
   void systemNameShouldComeFromTestClassViaJunitResource() {
